First take on Types
